Can set a sample in the wavetable

+++ src/soyboy/event.rs
@@ -3,7 +3,7 @@ pub enum Event {
     NoteOff { note: u16 },
     PitchBend { ratio: f64 },
     SweepReset { freq: f64 },
-    SetWaveTable { idx: usize, value: f64 },
+    SetWaveTable { idx: usize, value: i8 },
     ResetWaveTableAsSine,
     ResetWaveTableAtRandom,
 }
@@ -20,7 +20,7 @@ impl TryFrom<u32> for Event {
             1 => Ok(Event::NoteOff { note: 0 }),
             2 => Ok(Event::PitchBend { ratio: 0.0 }),
             3 => Ok(Event::SweepReset { freq: 0.0 }),
-            4 => Ok(Event::SetWaveTable { idx: 0, value: 0.0 }),
+            4 => Ok(Event::SetWaveTable { idx: 0, value: 0 }),
             5 => Ok(Event::ResetWaveTableAsSine),
             6 => Ok(Event::ResetWaveTableAtRandom),
             _ => Err(()),
